requirements.txt
cloudsync
cloudsync_onedrive
pytest
python-dotenv==0.14.0